+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 * feature#226: Panel Devices - add what devices are down and fail date
+* feature#228: Add Busiest panel - top utilized ports
+
 
 --- 4.0.2 ---